requirements.txt
cmake>=3.15
cmake-generators
GitPython
future_fstrings
pybind11
